% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqopr.R
\name{seqopr}
\alias{seqopr}
\alias{seqss}
\alias{lr.inf}
\title{Compute Operating Characteristics for Group Sequential Designs}
\usage{
seqopr(
  acc.per,
  acc.rate,
  control.rate = NULL,
  pct.imp = NULL,
  a.times,
  prop.strat = rep(1, length(control.rate)),
  alpha = 0.025,
  alphal = 0,
  p.con = 0.5,
  use = 6,
  usel = 6,
  oftr = alpha/50,
  oftrl = alphal/50,
  control.med = NULL,
  pct.reduc = NULL
)
}
\arguments{
\item{acc.per}{In \code{sequse} and \code{lr.inf}, the planned accrual
period for the study.  In \code{seqss}, a vector of length 2 giving the min
and max values of the range of possible accrual periods over which to
search.}

\item{acc.rate}{Expected accrual rate}

\item{control.rate}{Vector giving the hazard rate on the control treatment
in each stratum}

\item{pct.imp}{The percent improvement in the median failure time for the
experimental treatment over the control}

\item{a.times}{Planned chronological times of interim analyses.  Must be
positive and in increasing order.}

\item{prop.strat}{Vector giving the proportion of the total sample in each
stratum.  The vector is renormalized to sum to 1, so the values only need to
be proportional to the proportions.}

\item{alpha}{The one-sided significance level of the group sequential test}

\item{alphal}{The one-sided significance level used in the RCI monitoring
for stopping in favor of the null.  The confidence level of the RCI is
\code{1-2*alphal}.  If \code{alphal <= 0}, only the upper boundary is used.}

\item{p.con}{The proportion of subjects to be randomized to the control
group}

\item{use}{the type of use function: 1=O'Brien-Fleming, 2=Pocock, 3=linear,
4=one and a half, 5=quadratic, 6=truncated O'Brien-Fleming}

\item{usel}{The use function for determining critical values for the RCI
lower boundary (same codes as \code{use})}

\item{oftr}{The significance level at which the truncated O-F boundary is
truncated (upper boundary)}

\item{oftrl}{The significance level at which the truncated O-F boundary is
truncated (RCI lower boundary)}

\item{control.med}{Vector giving the median failure times on the control
treatment in each stratum (may be given instead of \code{control.rate})}

\item{pct.reduc}{The percent reduction in the hazard rate for the
experimental treatment relative to the control (may be given instead of
\code{pct.imp})}

\item{power}{The desired power for the study}

\item{add.fu}{The additional follow-up planned after the end of accrual
before the final analysis will be conducted}

\item{anal.int}{Interim analyses are planned to take place every
\code{anal.int} time units on the chronological time scale, beginning at
\code{first.anal}}

\item{first.anal}{Time of the first planned interim analysis}
}
\value{
\code{seqopr} returns a list with the following components (except
\code{acc.per}).  \code{seqss} returns a list with the following components
computed from the final design.  \item{Expected.inf }{A matrix giving the
expected information times at the planned analysis times and the expected
number of failures under the null and alternative and the sample size at
each analysis.} \item{Boundaries }{A matrix giving the boundaries on the
standard normal test statistic scale at the expected information at the
planned analysis times, plus columns giving the upper and lower boundary
crossing probabilities under H0 and H1 at each analysis} \item{Rej.Probs }{A
vector of 4 values giving the size and power (upper boundary crossing
probabilities under H0 and H1) and the total lower boundary crossing
probabilities under H0 and H1. } \item{stop.times }{A 2 x 4 matrix of
expected stopping times} \item{eta }{The Brownian motion mean parameter}
\item{call }{The call to \code{seqopr}} \item{acc.per}{The accrual period
that gives the desired power}

\code{lr.inf} returns a vector giving the values of the Brownian motion mean
parameter \code{eta} and the expected number of failures under the
alternative.
}
\description{
Computes operating characteristics (seqopr) and sample size (seqss) for
group sequential logrank tests in the setting of phase III studies with
failure time endpoints.
}
\details{
Assumes accrual is uniform at the rate \code{acc.rate} for \code{acc.per}
units of time, so the total number of subjects is \code{acc.rate * acc.per},
with interim analyses to be performed at the times in \code{a.times}.  The
final analysis is assumed to take place at the final time in \code{a.times}.
Significance levels at the analyses are determined from the error spending
rate function approach (see \code{sequse} for more details).  Optionally, an
asymmetric lower boundary for early stopping in favor of the null can be
included by specifying a value of \code{alphal > 0}.  An approximate lower
boundary based on repeated confidence intervals is used (see \code{sequse}).
Failure times are assumed to be exponentially distributed, with the
exponential failure hazard rates specified through the \code{control.rate}
parameter, or alternately the \code{control.med} parameter.  A stratified
sample with different failure rates in the strata can be specified by giving
a vector of control group hazard rates corresponding to the different
strata, and specifying the stratum proportions in \code{prop.strata}.  These
functions assume proportional hazards alternatives.  The improvement from
using the experimental treatment instead of control is specified as the
PERCENT improvement in the median time to failure.  This is related to the
hazard ratio through hazard ratio = \code{1/(1+pct.imp/100)}.  Alternately,
the percent reduction in the hazard rate (\code{pct.reduc}) can be
specified.

Given the specifications above, \code{seqopr} computes the boundaries,
power, boundary crossing probabilities and expected stopping times.  The
expected stopping times under the null are computed using the specified
calendar times of analysis and the null failure rates, but using the
information times and boundaries based on the expected information at these
times as determined under the alternative.  They therefore do not typically
give the correct expected stopping times under the design as it will be
implemented in practice.

In \code{seqss}, the desired power is specified, and a range is given for
the accrual period.  The program then searches for the accrual period that
gives the desired power.  The timing of the final analysis after the
completion of accrual (parameter \code{add.fu}) is kept fixed.  Since the
study duration is variable, an approximate interim analysis schedule is
specified through the parameters \code{anal.int} and \code{first.anal}.
When early stopping in favor of the null is included, it is possible that
some designs examined in the search will have incompatible specifications.
If this happens, try reducing the upper limit on \code{acc.per}.  If the
range of \code{acc.per} specified does not include a solution, then the
algorithm should give an error message stating that the values at the end
points are not of opposite signs.  If this occurs, try increasing the width
of the interval.

Given accrual and follow-up information and failure rates, \code{lr.inf}
computes the total planned information and the Brownian motion mean
parameter \code{eta}.  The main use of this function is to get the value of
\code{eta} for computing the lower boundary in \code{sequse}.
}
\examples{
seqopr(3, 200, 0.1, 50, c(2, 3, 4, 5))
seqss(0.8, c(2, 6), 200, 2, 0.5, 2, 0.1, 50)
lr.inf(3, 200, 2, 0.1, 50)

}
\seealso{
\code{\link{sequse}}; \code{\link{powlgrnk6}}; \code{\link{seqp}}
}
\keyword{design}
