% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqp.R
\name{seqp}
\alias{seqp}
\title{Compute Boundary Crossing Probabilities}
\usage{
seqp(inf.times, upper, lower = NULL, eta = 0)
}
\arguments{
\item{inf.times}{Information times of analyses}

\item{upper}{Upper boundary (for rejecting H0) on standard normal scale}

\item{lower}{Lower boundary (for rejecting HA) on standard normal scale}

\item{eta}{The mean parameter on the Brownian motion process scale.}
}
\value{
A \code{length(inf.times)} by 5 matrix with columns giving the
information times, the upper boundary, the lower boundary, the probabilities
of crossing the upper boundary, and the probabilities of crossing the lower
boundary.
}
\description{
Computes probabilities of crossing arbitrary user specified upper and/or
lower boundaries
}
\details{
For a group sequential design with arbitrary boundaries, computes the
probability of crossing the upper and lower boundary at each analysis.  The
probability at each analysis is the probability of not crossing either
boundary at earlier analyses and crossing at the current analysis.

\code{eta} should be 0 to compute probabilites under the null, and should be
set to a value which is a function of the alternative on a transformed scale
to compute probabilties under the alternative.  For survival studies, eta
can be computed using the functions \code{seqopr} or \code{lr.inf}.
(\code{eta} is the mean of Y(1), where Y(t) is the partial sum process
standardized to have variance 1 at t=1.)
}
\examples{
seqp((1:4)/4,c(2,2,3,2),eta=3)
seqp((1:4)/4,c(2,2,3,2),eta=0)

}
\seealso{
\code{\link{sequse}}, \code{\link{seqopr}}, \code{\link{lr.inf}}
}
\keyword{design}
