% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequse.R
\name{sequse}
\alias{sequse}
\title{Compute Group Sequential Use Function Boundaries}
\usage{
sequse(inf, alpha = .025, use = 6, eta = 0, alphal = 0,
       usel = 6, oftr = alpha / 50, oftrl = alphal / 50)
}
\arguments{
\item{inf}{information times of analyses (length <= 30); must be positive,
increasing and <= 1}

\item{alpha}{one-sided significance level of the group sequential test}

\item{use}{the type of use function: 1=O'Brien-Fleming, 2=Pocock, 3=linear,
4=one and a half, 5=quadratic, 6=truncated O'Brien-Fleming}

\item{eta}{The mean parameter on the Brownian motion process scale.  Only
needed for the RCI lower boundary.}

\item{alphal}{The one-sided significance level used in the RCI monitoring
for stopping in favor of the null.  The confidence level of the RCI is
\code{1-2*alphal}.  If \code{alphal <= 0}, only the upper boundary is
computed.}

\item{usel}{The use function for determining critical values for the RCI
lower boundary (same codes as \code{use})}

\item{oftr}{The significance level at which the truncated O-F boundary is
truncated (upper boundary)}

\item{oftrl}{The significance level at which the truncated O-F boundary is
truncated (RCI lower boundary)}
}
\value{
a matrix with \code{length(inf)} rows giving the critical values on
the standard normal scale at the specified information times.  If
\code{alphal<=0}, there is a single column giving the critical values for
the upper one-sided boundary.  If \code{alphal>0}, then there is a second
column giving the critical value for the lower boundary for early stopping
in favor of the null.  Note that these critical values are on the normalized
test statistic scale, which are NOT the critical values used in constructing
the RCI.
}
\description{
Computes the one-sided group sequential boundary for a specified use
function at specified analysis times.  Optionally, also computes an
asymmetric lower boundary based on repeated confidence interval (RCI)
monitoring for early stopping in favor of the null hypothesis.
}
\details{
Calculates the group sequential boundaries for repeated significance tests
in group sequential analysis of clinical trials.  The algorithm is based on
the use function approach proposed by Lan and DeMets (1983, Biometrika) and
investigated further by Kim and DeMets (1987, Biometrika).  The information
time corresponds to the proportion of statistical information, which is the
essentially the same as the number of failures in a proportional hazards
model for failure time endpoints.

For the truncated O-F boundary (\code{use=6}), first the regular O-F
boundary is computed, but if the critical value is larger than the specified
truncation value, the truncated value is used instead.  The actual error
spent is computed, and the early over-spending is made up as quickly as
possible, after which the boundary is similar to the ordinary O-F boundary.

With asymmetric monitoring, the study is stopped early in favor of the null
hypothesis if a `lower boundary' is crossed.  Here the lower boundary is
based on a repeated confidence interval on the log hazard ratio (see
Jennison and Turnbull, 1990).  The RCI is constructed using the critical
value from the one-sided boundary specified by \code{usel} and
\code{alphal}.  In the program, this is converted to a boundary on the
logrank statistic.  This requires information on the alternative and the
total planned information for the study, which is specified through the
parameter \code{eta}.  The value of \code{eta} can be obtained from the
functions \code{seqopr} and \code{lr.inf}.  Introducing a lower boundary
reduces the probability of crossing the upper boundary.  After determining
the lower boundary, the program computes the upper boundary taking into
account the lower boundary.  The upper boundary will thus be affected by the
parameters specified for the lower boundary.  It is quite easy to specify
incompatible combinations for the upper and lower boundary, especially if
\code{alpha} or \code{alphal} is very large or if \code{eta} is small.
}
\note{
Interface to the Fortran code for the program \code{sequse}, which was
written by Kyungmann Kim and modified for truncated O-F boundaries and
asymmetric lower boundaries by Bob Gray
}
\examples{
sequse((1:4)/4)
sequse((1:4)/4, use = 6)
sequse((1:4)/4, use = 6, alphal = .025, eta = 2)

}
\references{
Lan and DeMets (1983). \emph{Biometrika}.

Kim and DeMets (1987). \emph{Biometrika}.

Jennison and Turnbull (1990). \emph{Statistical Science} \strong{5}:299-317.
}
\seealso{
\code{\link{seqopr}}, \code{\link{lr.inf}}, \code{\link{seqp}}
}
\keyword{design}
